requirements.txt
aiogram==3.26.0
python-dotenv==1.0.1
psutil==6.1.1
pyyaml==6.0.2
